// backend/services/libs/shared/src/enum/projectProposalStage.enum.ts
export enum ProjectProposalStage {
  SUBMITTED_INF = "SUBMITTED_INF",
  APPROVED_INF = "APPROVED_INF",
  REJECTED_INF = "REJECTED_INF",
  SUBMITTED_COST_QUOTATION = "SUBMITTED_COST_QUOTATION",
  SUBMITTED_PROPOSAL = "SUBMITTED_PROPOSAL",
  SUBMITTED_VALIDATION_AGREEMENT = "SUBMITTED_VALIDATION_AGREEMENT",
  ACCEPTED_PROPOSAL = "ACCEPTED_PROPOSAL",
  REJECTED_PROPOSAL = "REJECTED_PROPOSAL",
  SUBMITTED_CMA = "SUBMITTED_CMA",
  REJECTED_CMA = "REJECTED_CMA",
  APPROVED_CMA = "APPROVED_CMA",
  VALIDATION_PENDING = "VALIDATION_PENDING",
  REJECTED_VALIDATION = "REJECTED_VALIDATION",
  AUTHORISED = "AUTHORISED",

  PENDING = "PENDING",
  REJECTED = "REJECTED",
  APPROVED = "APPROVED",
  PDD_SUBMITTED = "PDD_SUBMITTED",
  PDD_REJECTED_BY_CERTIFIER = "PDD_REJECTED_BY_CERTIFIER",
  PDD_APPROVED_BY_CERTIFIER = "PDD_APPROVED_BY_CERTIFIER",
  PDD_REJECTED_BY_DNA = "PDD_REJECTED_BY_DNA",
  PDD_APPROVED_BY_DNA = "PDD_APPROVED_BY_DNA",
  VALIDATION_REPORT_SUBMITTED = "VALIDATION_REPORT_SUBMITTED",
  VALIDATION_DNA_REJECTED = "VALIDATION_REPORT_REJECTED",
  AUTHORIZED = "AUTHORIZED",
}